requirements.txt
flask
numpy
scipy.stats
yfinance